refactor: migrate proposal page to dedicated content component
